# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Atzicay API",
        "description": "API documentation for Atzicay",
        "contact": {
            "email": "atzicay@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "api/atzicay/v1/"
        }
    ],
    "paths": {
...
        "/words/{id}": {
            "get": {
                "tags": [
                    "Words"
                ],
                "summary": "Get Word by ID",
                "description": "Retrieves a Word by its ID.",
                "operationId": "66f459fbb42059cb691b84a75c470a63",
                "parameters": [
                    {
                        "name": "Id",
                        "in": "path",
                        "description": "ID of the Word to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Word found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Words"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Word not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Words"
                ],
                "summary": "Update a Word",
                "description": "Updates an existing Word.",
                "operationId": "4122ed1adaaf5dce0f8da48e2f2a11ff",
                "parameters": [
                    {
                        "name": "Id",
                        "in": "path",
                        "description": "ID of the Word to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WordsDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Word updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Words"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/words": {
            "post": {
                "tags": [
                    "Words"
                ],
                "summary": "Create a new Word",
                "description": "Creates a new Word.",
                "operationId": "7d038c99b0b628abfdd04d0e548ebb74",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WordsDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Word created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Words"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "WordsDTO": {
                "title": "WordsDTO",
                "description": "Data Transfer Object for Words",
                "required": [
                    "SolveTheWordId",
                    "Word",
                    "Orientation"
                ],
                "properties": {
                    "SolveTheWordId": {
                        "description": "The ID of the word to solve",
                        "type": "integer"
                    },
                    "Word": {
                        "description": "The word to be solved",
                        "type": "string"
                    },
                    "Orientation": {
                        "description": "The orientation of the word",
                        "type": "string",
                        "enum": [
                            "HL",
                            "HR",
                            "VU",
                            "VD",
                            "DU",
                            "DD"
                        ]
                    }
                },
                "type": "object"
            },
...
            "Words": {
                "title": "Words",
                "description": "Words entity schema",
                "properties": {
                    "Id": {
                        "description": "Primary key of the Words entity",
                        "type": "integer"
                    },
                    "SolveTheWordId": {
                        "description": "Foreign key referencing SolveTheWord entity",
                        "type": "integer"
                    },
                    "Word": {
                        "description": "The word itself",
                        "type": "string"
                    },
                    "Orientation": {
                        "description": "Orientation of the word",
                        "type": "string",
                        "enum": [
                            "HL",
                            "HR",
                            "VU",
                            "VD",
                            "DU",
                            "DD"
                        ]
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}